get todo with its state
load todos based on it's state.
Finished todos are crossed in the todo list after loading the page. These todos can be found in the completed list.

Added a check icon for each finished todo in the completed list, click this icon will trigger a delete request to the json-server